requirements.inference.txt
# Minimal dependencies for inference and CLI gameplay
torch==2.8.0
chess==1.10.0
pyyaml==6.0.1
numpy==2.1.3
fastapi==0.115.0
uvicorn==0.30.6

# Optional for rendering nice boards
# matplotlib==3.9.0
# cairosvg==2.7.1
# Pillow==10.4.0

